% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erfgoed_theme.R
\name{bcomma}
\alias{bcomma}
\title{wrapper to set defaults for scales::format_format}
\usage{
bcomma(x)
}
\arguments{
\item{x}{Vector to be formatted.
returns a vector formatted with comma as a decimal seperator}
}
\description{
wrapper to set defaults for scales::format_format
}

