% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erfgoed_kleuren.R
\name{erfgoed_palette}
\alias{erfgoed_palette}
\title{colorpalette}
\usage{
erfgoed_palette(n)
}
\arguments{
\item{n}{number of colours (integer)}
}
\value{
colour palette
}
\description{
colorpalette
}

