% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erfgoed_theme.R
\name{oe_pdf}
\alias{oe_pdf}
\title{Function to return a default pdf graphics device}
\usage{
oe_pdf(file = "output.pdf", width = 9, height = 7)
}
\arguments{
\item{file}{Output filename}

\item{width}{Defaults to 9}

\item{height}{Defaults to 7}
}
\value{
A pdf graphics device
}
\description{
Function to return a default pdf graphics device
}

