% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erfgoed_theme.R
\name{oe_png}
\alias{oe_png}
\title{Function to return a default png graphics device}
\usage{
oe_png(file = "output.png", width = 6.3, height = 5.9, units = "in",
  res = 600)
}
\arguments{
\item{file}{Output filename}

\item{width}{Defaults to 6.3}

\item{height}{Defaults to 5.9}

\item{units}{defaults to inch}

\item{res}{defaults to 600}
}
\value{
A png graphics device
}
\description{
Function to return a default png graphics device
}

